Competed the Advanced Authentication & Authorization

+++ frontend/src/api/axios.js
@@ -4,6 +4,7 @@ const API = axios.create({
   baseURL: "http://127.0.0.1:8000",
 });
 
+// 🔹 Attach access token
 API.interceptors.request.use((req) => {
   const token = localStorage.getItem("token");
   if (token) {
@@ -12,13 +13,49 @@ API.interceptors.request.use((req) => {
   return req;
 });
 
+// 🔹 Handle 401 with refresh logic
 API.interceptors.response.use(
   (response) => response,
-  (error) => {
-    if (error.response && error.response.status === 401) {
-      localStorage.clear()
-      window.location.href = "/";
+  async (error) => {
+    const originalRequest = error.config;
+
+    if (
+      error.response &&
+      error.response.status === 401 &&
+      !originalRequest._retry
+    ) {
+      originalRequest._retry = true;
+
+      const refreshToken = localStorage.getItem("refresh_token");
+
+      // if no refresh token → logout
+      if (!refreshToken) {
+        localStorage.clear();
+        window.location.href = "/";
+        return Promise.reject(error);
+      }
+
+      try {
+        const res = await axios.post(
+          "http://127.0.0.1:8000/auth/refresh",
+          {
+            refresh_token: refreshToken,
+          }
+        );
+        const newAccessToken = res.data.access_token;
+
+        localStorage.setItem("token", newAccessToken);
+
+        originalRequest.headers.Authorization = `Bearer ${newAccessToken}`;
+
+        return API(originalRequest);
+      } catch (err) {
+        localStorage.clear();
+        window.location.href = "/";
+        return Promise.reject(err);
+      }
     }
+
     return Promise.reject(error);
   }
 );
